// src/types/IArtworks.ts
export interface IArtwork {
    alt_artist_ids: string[];
    alt_classification_ids: string[];
    alt_image_ids: string[];
    alt_material_ids: string[];
    alt_style_ids: string[];
    alt_subject_ids: string[];
    alt_technique_ids: string[];
    alt_titles: null;
    api_link: string;
    api_model: string;
    artist_display: string;
    artist_id: number;
    artist_ids: number[];
    artist_title: string;
    artist_titles: string[];
    artwork_type_id: number;
    artwork_type_title: string;
    boost_rank: null;
    catalogue_display: null;
    category_ids: string[];
    category_titles: string[];
    classification_id: string;
    classification_ids: string[];
    classification_title: string;
    classification_titles: string[];
    color: {
        h: number;
        l: number;
        s: number;
        percentage: number;
        population: number;
    };
    colorfulness: number;
    copyright_notice: string;
    credit_line: string;
    date_display: string;
    date_end: number;
    date_qualifier_id: null;
    date_qualifier_title: string;
    date_start: number;
    department_id: string;
    department_title: string;
    description: string;
    dimensions: string;
    dimensions_detail: any[]; 
    document_ids: string[];
    edition: null;
    exhibition_history: string;
    fiscal_year: number;
    fiscal_year_deaccession: null;
    gallery_id: null;
    gallery_title: null;
    has_advanced_imaging: boolean;
    has_educational_resources: boolean;
    has_multimedia_resources: boolean;
    has_not_been_viewed_much: boolean;
    id: number;
    imageUrl: string;
    image_id: string;
    inscriptions: string;
    internal_department_id: number;
    is_boosted: boolean;
    is_on_view: boolean;
    is_public_domain: boolean;
    is_zoomable: boolean;
    latitude: null;
    latlon: null;
    longitude: null;
    main_reference_number: string;
    material_id: string;
    material_ids: string[];
    material_titles: string[];
    max_zoom_window_size: number;
    medium_display: string;
    nomisma_id: null;
    on_loan_display: string;
    place_of_origin: string;
    provenance_text: string;
    publication_history: string;
    publishing_verification_level: string;
    section_ids: number[];
    section_titles: string[];
    short_description: string;
    site_ids: string[];
    sound_ids: string[];
    source_updated_at: string;
    style_id: null;
    style_ids: string[];
    style_title: null;
    style_titles: string[];
    subject_id: string;
    subject_ids: string[];
    subject_titles: string[];
    suggest_autocomplete_all: any[]; 
    suggest_autocomplete_boosted: string;
    technique_id: string;
    technique_ids: string[];
    technique_titles: string[];
    term_titles: string[];
    text_ids: string[];
    theme_titles: string[];
    thumbnail: {
        lqip: string;
        width: number;
        height: number;
        alt_text: string;
    };
    timestamp: string;
    title: string;
    updated_at: string;
    video_ids: string[];
}